script: introduce a new txouttype.h
solver.h is included by policy.h only to get access to the TxoutType
enum. Separate it in its own module to let us strip solver.h from
the kernel headers.
