+++ src/script/txouttype.h
@@ -1,0 +1,28 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2023 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_SCRIPT_TXOUTTYPE_H
+#define BITCOIN_SCRIPT_TXOUTTYPE_H
+
+#include <string>
+
+enum class TxoutType {
+    NONSTANDARD,
+    // 'standard' transaction types:
+    PUBKEY,
+    PUBKEYHASH,
+    SCRIPTHASH,
+    MULTISIG,
+    NULL_DATA, //!< unspendable OP_RETURN script that carries data
+    WITNESS_V0_SCRIPTHASH,
+    WITNESS_V0_KEYHASH,
+    WITNESS_V1_TAPROOT,
+    WITNESS_UNKNOWN, //!< Only for Witness versions not already defined above
+};
+
+/** Get the name of a TxoutType as a string */
+std::string GetTxnOutputType(TxoutType t);
+
+#endif // BITCOIN_SCRIPT_TXOUTTYPE_H
